fix: invalid package cyclic dependency
